Add files via upload
This is the first release, may be subject to modifications and enhancements. This first release try to maintain the first logic of the project, the use of linked list to store info about the multiple files that the program can run, each one will have his respectively node with the data about the FD number, last position readed in the previous run, the BUFFER_SIZE that need to be defined at the moment of compilation with the flag: "-D BUFFER_SIZE=n", being 'n' the desired size, and a pointer to the next node. Each time the programs read a new file, it will create a new node at the beginning of the list, the program will also transverse the current list in order to find a node with the same fd as the input, if it finds will gather the info, update the buffer_size and re run the program with the new parameters from the previous index position from the first run. More explanations and improvements are expected.